requirements.txt
pyspark==3.5.1
flake8==6.0.0
boto3==1.35.70
jinja2==3.1.2
PyYAML==6.0.1
pydeequ==1.1.1
delta-spark==3.1.0
# great-expectations==0.18.8
pycountry==24.6.1
pyarrow==19.0.0
